+++ changelog.md
@@ -24,6 +24,7 @@
 - Fix arms on posts dropping blocks when broken (BluSunrize)
 - Fix crate minecarts crashing when inserted into with a hopper (BluSunrize)
 - Fix barrel minecarts crashing when right-clicked (BluSunrize)
+- Fix villagers trying to pathfind through multiblocks (BluSunrize)
 - Translations Added/Updated: de_de (AriaElidove)
 
 ##### Version 1.20.4-11.5.0-179